refactor: rename adm.py to provider.py
