+++ merino/providers/__init__.py
@@ -7,9 +7,9 @@ from timeit import default_timer as timer
 from merino import metrics
 from merino.config import settings
 from merino.exceptions import InvalidProviderError
-from merino.providers.adm.adm import Provider as AdmProvider
-from merino.providers.adm.adm import TestBackend
 from merino.providers.adm.backends.remotesettings import LiveBackend
+from merino.providers.adm.provider import Provider as AdmProvider
+from merino.providers.adm.provider import TestBackend
 from merino.providers.base import BaseProvider
 from merino.providers.top_picks import Provider as TopPicksProvider
 from merino.providers.weather.backends.accuweather import AccuweatherBackend
